Move `pathOr`, `queryParams` utilities
