+++ packages/manager/src/features/Linodes/LinodeCreate/utilities.ts
@@ -1,5 +1,8 @@
 import { omitProps } from '@linode/ui';
-import { isNotNullOrUndefined } from '@linode/utilities';
+import {
+  getQueryParamsFromQueryString,
+  isNotNullOrUndefined,
+} from '@linode/utilities';
 import { enqueueSnackbar } from 'notistack';
 import { useCallback } from 'react';
 import { useHistory } from 'react-router-dom';
@@ -11,7 +14,6 @@ import { sendCreateLinodeEvent } from 'src/utilities/analytics/customEventAnalyt
 import { sendLinodeCreateFormErrorEvent } from 'src/utilities/analytics/formEventAnalytics';
 import { isPrivateIP } from 'src/utilities/ipUtils';
 import { utoa } from 'src/utilities/metadata';
-import { getQueryParamsFromQueryString } from 'src/utilities/queryParams';
 
 import {
   getLegacyInterfaceFromLinodeInterface,
