+++ packages/manager/src/features/Linodes/LinodesDetail/LinodesDetail.tsx
@@ -1,4 +1,5 @@
 import { CircleProgress, ErrorState } from '@linode/ui';
+import { getQueryParamsFromQueryString } from '@linode/utilities';
 import { createLazyRoute } from '@tanstack/react-router';
 import * as React from 'react';
 import {
@@ -12,7 +13,6 @@ import {
 
 import { SuspenseLoader } from 'src/components/SuspenseLoader';
 import { useLinodeQuery } from 'src/queries/linodes/linodes';
-import { getQueryParamsFromQueryString } from 'src/utilities/queryParams';
 
 import type { LinodeConfigAndDiskQueryParams } from 'src/features/Linodes/types';
