+++ packages/manager/src/features/Linodes/types.ts
@@ -1,5 +1,5 @@
 import type { LinodeCreateType } from './LinodeCreate/types';
-import type { BaseQueryParams } from 'src/utilities/queryParams';
+import type { BaseQueryParams } from '@linode/utilities';
 
 export type DialogType =
   | 'delete'
