+++ packages/manager/src/features/TopMenu/SearchBar/SearchBar.tsx
@@ -1,5 +1,8 @@
 import { Autocomplete, Box, IconButton, TextField } from '@linode/ui';
-import { isNotNullOrUndefined } from '@linode/utilities';
+import {
+  getQueryParamsFromQueryString,
+  isNotNullOrUndefined,
+} from '@linode/utilities';
 import Close from '@mui/icons-material/Close';
 import { useMediaQuery, useTheme } from '@mui/material';
 import * as React from 'react';
@@ -27,7 +30,6 @@ import { formatLinode } from 'src/store/selectors/getSearchEntities';
 import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
 import { extendTypesQueryResult } from 'src/utilities/extendType';
 import { isNilOrEmpty } from 'src/utilities/isNilOrEmpty';
-import { getQueryParamsFromQueryString } from 'src/utilities/queryParams';
 
 import { StyledIconButton, StyledSearchIcon } from './SearchBar.styles';
 import { SearchSuggestion } from './SearchSuggestion';
