+++ packages/manager/src/features/VPCs/VPCCreate/FormComponents/SubnetContent.tsx
@@ -1,11 +1,11 @@
 import { Notice } from '@linode/ui';
+import { getQueryParamsFromQueryString } from '@linode/utilities';
 import * as React from 'react';
 import { useFormContext } from 'react-hook-form';
 import { useLocation } from 'react-router-dom';
 
 import { Link } from 'src/components/Link';
 import { sendLinodeCreateFormInputEvent } from 'src/utilities/analytics/formEventAnalytics';
-import { getQueryParamsFromQueryString } from 'src/utilities/queryParams';
 
 import { VPC_CREATE_FORM_SUBNET_HELPER_TEXT } from '../../constants';
 import { MultipleSubnetInput } from '../MultipleSubnetInput';
