+++ packages/manager/src/hooks/useCreateVPC.ts
@@ -1,5 +1,6 @@
 import { yupResolver } from '@hookform/resolvers/yup';
 import { isEmpty } from '@linode/api-v4';
+import { getQueryParamsFromQueryString } from '@linode/utilities';
 import { createVPCSchema } from '@linode/validation';
 import * as React from 'react';
 import { useForm } from 'react-hook-form';
@@ -9,7 +10,6 @@ import { useGrants, useProfile } from 'src/queries/profile/profile';
 import { useRegionsQuery } from 'src/queries/regions/regions';
 import { useCreateVPCMutation } from 'src/queries/vpcs/vpcs';
 import { sendLinodeCreateFormStepEvent } from 'src/utilities/analytics/formEventAnalytics';
-import { getQueryParamsFromQueryString } from 'src/utilities/queryParams';
 import { scrollErrorIntoView } from 'src/utilities/scrollErrorIntoView';
 import { DEFAULT_SUBNET_IPV4_VALUE } from 'src/utilities/subnets';
