+++ packages/manager/src/layouts/OAuthCallback.tsx
@@ -1,4 +1,4 @@
-import { capitalize } from '@linode/utilities';
+import { capitalize, getQueryParamsFromQueryString } from '@linode/utilities';
 import React from 'react';
 import { useHistory, useLocation } from 'react-router-dom';
 
@@ -10,7 +10,6 @@ import {
   getAuthCode,
   setAuthToken,
 } from 'src/utilities/authentication';
-import { getQueryParamsFromQueryString } from 'src/utilities/queryParams';
 import { getEnvLocalStorageOverrides } from 'src/utilities/storage';
 
 const localStorageOverrides = getEnvLocalStorageOverrides();
