+++ packages/utilities/src/helpers/index.ts
@@ -18,3 +18,5 @@ export * from './pluralize';
 export * from './minute-conversion';
 export * from './mockLocalStorage';
 export * from './nullOrUndefined';
+export * from './pathOr';
+export * from './queryParams';
