+++ packages/utilities/src/helpers/queryParams.test.ts
@@ -1,3 +1,5 @@
+import { describe, expect, it } from 'vitest';
+
 import {
   getQueryParamFromQueryString,
   getQueryParamsFromQueryString,
